# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "description": "NIFTY intraday trajectory pattern API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/history/build": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Build History",
        "description": "Read NIFTY/NSE/1m source bars and persist only project-owned derived records.\n\nRanges are inclusive and bounded to 366 calendar days per request; use\nhistory/sync to process all history without manual ranges. Rebuilding\nreplaces derived records and processed-day flags for this configuration/\ndate range, never source records. Incomplete/invalid sessions are reported\nand flagged, not repaired. E days are retained. Dates with no source rows\nare not assumed to be trading sessions. CLI: `build`.",
        "operationId": "buildHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/history/sync": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Sync History",
        "description": "Process every source session not yet flagged for this configuration.\n\nThe first sync processes all available history; later syncs process only\nnew or gap dates, plus excluded dates among the three most recent sessions.\nOnly sessions through the last complete session (15:30 IST) are processed.\nrebuild=true deletes this configuration's derived records and flags, then\nreprocesses everything. Each batch commits separately, so an interrupted\nsync resumes on the next call. Long-running on a first load. CLI: `sync`.",
        "operationId": "syncHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/history/status": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "History Status",
        "description": "Report stored coverage and processed-day flags for a built configuration. CLI: `status`.",
        "operationId": "historyStatus",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64,
              "pattern": "^[0-9a-f]{64}$",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryStatus"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/patterns/match": {
      "post": {
        "tags": [
          "patterns"
        ],
        "summary": "Match Pattern",
        "description": "Match completed hourly tokens using strictly earlier historical dates.\n\nSupply bars, or omit them to load the requested developing day from the\napproved source. Backoff never returns a distribution with fewer than N\nmembers. Outcome windows always start at the requested hour, not the\nshorter selected prefix. Position 7 is never a matching key. CLI: `match`.",
        "operationId": "matchPattern",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/evaluation/walk-forward": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Evaluate",
        "description": "Evaluate built history chronologically against all-prior-days outcomes.\n\nReturns direction hit rate, three-class Brier score, P(Up) calibration,\nempirical P10\u2013P90 coverage and support/backoff by hour. Conditional and\nbaseline comparisons use the same supported test days. No evaluation\nwrites, time weighting, fitted ML, or future-day training occurs. CLI: `evaluate`.",
        "operationId": "evaluateWalkForward",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PatternConfig"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "config",
          "start_date",
          "end_date"
        ],
        "title": "BuildRequest"
      },
      "BuildResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "built",
            "title": "Status"
          },
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "source_days": {
            "type": "integer",
            "title": "Source Days"
          },
          "stored_days": {
            "type": "integer",
            "title": "Stored Days"
          },
          "unique_patterns": {
            "type": "integer",
            "title": "Unique Patterns"
          },
          "excluded_sessions": {
            "items": {
              "$ref": "#/components/schemas/ExcludedSession"
            },
            "type": "array",
            "title": "Excluded Sessions"
          },
          "created_project_storage": {
            "type": "boolean",
            "title": "Created Project Storage"
          }
        },
        "type": "object",
        "required": [
          "status",
          "config_id",
          "source",
          "start_date",
          "end_date",
          "source_days",
          "stored_days",
          "unique_patterns",
          "excluded_sessions",
          "created_project_storage"
        ],
        "title": "BuildResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "details"
        ],
        "title": "ErrorResponse"
      },
...
      "EvaluationRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PatternConfig"
          },
          "support_target": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Support Target"
          },
          "history_start_date": {
            "type": "string",
            "format": "date",
            "title": "History Start Date"
          },
          "test_start_date": {
            "type": "string",
            "format": "date",
            "title": "Test Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "positions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 1,
            "title": "Positions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "config",
          "support_target",
          "history_start_date",
          "test_start_date",
          "end_date"
        ],
        "title": "EvaluationRequest"
      },
      "EvaluationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "evaluated",
              "no_test_days"
            ],
            "title": "Status"
          },
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "support_target": {
            "type": "integer",
            "title": "Support Target"
          },
          "test_start_date": {
            "type": "string",
            "format": "date",
            "title": "Test Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "per_hour": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Per Hour"
          },
          "interpretation": {
            "type": "string",
            "title": "Interpretation"
          }
        },
        "type": "object",
        "required": [
          "status",
          "config_id",
          "method",
          "support_target",
          "test_start_date",
          "end_date",
          "per_hour",
          "interpretation"
        ],
        "title": "EvaluationResponse"
      },
      "ExcludedSession": {
        "properties": {
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "trade_date",
          "reason",
          "details"
        ],
        "title": "ExcludedSession"
      },
...
      "HistoryStatus": {
        "properties": {
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "stored_days": {
            "type": "integer",
            "title": "Stored Days"
          },
          "first_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Date"
          },
          "last_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Date"
          },
          "unique_patterns": {
            "type": "integer",
            "title": "Unique Patterns"
          },
          "processed_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Days"
          },
          "excluded_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Days"
          },
          "last_processed_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Processed Date"
          }
        },
        "type": "object",
        "required": [
          "config_id",
          "stored_days",
          "first_date",
          "last_date",
          "unique_patterns",
          "processed_days",
          "excluded_days",
          "last_processed_date"
        ],
        "title": "HistoryStatus"
      },
      "MatchRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PatternConfig"
          },
          "support_target": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Support Target"
          },
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "through_position": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Through Position"
          },
          "bars": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Bar"
                },
                "type": "array",
                "maxItems": 1000,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Bars"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "config",
          "support_target",
          "trade_date",
          "through_position"
        ],
        "title": "MatchRequest"
      },
...
      "PredictionResponse": {
        "properties": {
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "through_position": {
            "type": "integer",
            "title": "Through Position"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "requested_prefix": {
            "type": "string",
            "title": "Requested Prefix"
          },
          "requested_prefix_match_count": {
            "type": "integer",
            "title": "Requested Prefix Match Count"
          },
          "selected_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Prefix"
          },
          "match_count": {
            "type": "integer",
            "title": "Match Count"
          },
          "backoff_positions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backoff Positions"
          },
          "status": {
            "type": "string",
            "enum": [
              "estimated",
              "insufficient_support"
            ],
            "title": "Status"
          },
          "support_target": {
            "type": "integer",
            "title": "Support Target"
          },
          "estimate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Estimate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "trade_date",
          "config_id",
          "through_position",
          "time",
          "requested_prefix",
          "requested_prefix_match_count",
          "selected_prefix",
          "match_count",
          "backoff_positions",
          "status",
          "support_target",
          "estimate"
        ],
        "title": "PredictionResponse"
      },
...
      "SyncRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PatternConfig"
          },
          "rebuild": {
            "type": "boolean",
            "title": "Rebuild",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "config"
        ],
        "title": "SyncRequest"
      },
      "SyncResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "synced",
            "title": "Status"
          },
          "config_id": {
            "type": "string",
            "title": "Config Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "rebuild": {
            "type": "boolean",
            "title": "Rebuild"
          },
          "through_date": {
            "type": "string",
            "format": "date",
            "title": "Through Date"
          },
          "source_days": {
            "type": "integer",
            "title": "Source Days"
          },
          "already_processed_days": {
            "type": "integer",
            "title": "Already Processed Days"
          },
          "pending_days": {
            "type": "integer",
            "title": "Pending Days"
          },
          "stored_days": {
            "type": "integer",
            "title": "Stored Days"
          },
          "excluded_sessions": {
            "items": {
              "$ref": "#/components/schemas/ExcludedSession"
            },
            "type": "array",
            "title": "Excluded Sessions"
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/SyncBatch"
            },
            "type": "array",
            "title": "Batches"
          },
          "created_project_storage": {
            "type": "boolean",
            "title": "Created Project Storage"
          }
        },
        "type": "object",
        "required": [
          "status",
          "config_id",
          "source",
          "rebuild",
          "through_date",
          "source_days",
          "already_processed_days",
          "pending_days",
          "stored_days",
          "excluded_sessions",
          "batches",
          "created_project_storage"
        ],
        "title": "SyncResponse"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API path"
    }
  ]
}